github-ci: add build.yml workflow
adding build.yml workflow which is triggered on push and pull requests.
The workflow pulls sources from the manifest using git-repo tool, syncing a local mirror. This local mirror is used for caching and saves a lot of time in subsequent pipelines.
Similarly, openharmony prebuilts are also downloaded and cached into persistent storage.
Finally, a build is executed for the rk3568 target with ccache stored in persistent storage as well to speed up the builds.

Signed-off-by: Francesco Pham <francesco.pham@huawei.com>